asm/syreg: Switch to 64-bit register for config_sctlr_el1
/run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/sysreg.h:260:43: error: value size does not match register size specified by the constraint and modifier [-Werror,-Wasm-operand-widths]
  260 |         asm volatile("mrs %0, sctlr_el1" : "=r" (val));
      |                                                  ^
/run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/sysreg.h:260:20: note: use constraint modifier "w"
  260 |         asm volatile("mrs %0, sctlr_el1" : "=r" (val));
      |                           ^~
      |                           %w0
/run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/sysreg.h:263:44: error: value size does not match register size specified by the constraint and modifier [-Werror,-Wasm-operand-widths]
  263 |         asm volatile("msr sctlr_el1, %0" : : "r" (val));
      |                                                   ^
/run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/sysreg.h:263:31: note: use constraint modifier "w"
  263 |         asm volatile("msr sctlr_el1, %0" : : "r" (val));
      |                                      ^~
      |                                      %w0

<inline asm>:1:6: error: invalid operand for instruction
    1 |         mrs w8, sctlr_el1
      |             ^
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/kernel/armv8_deprecated.c:9:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/cpu.h:16:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/node.h:17:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/device.h:17:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/kobject.h:20:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/list.h:8:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/kernel.h:13:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/printk.h:8:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/include/linux/cache.h:5:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/cache.h:19:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/cachetype.h:19:
In file included from /run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/cputype.h:118:
/run/media/kevin/Daten/NLA/LineageOS21/kernel/nokia/qcom-msm8998/arch/arm64/include/asm/sysreg.h:263:15: error: invalid operand for instruction
  263 |         asm volatile("msr sctlr_el1, %w0" : : "r" (val));
      |                      ^
<inline asm>:1:17: note: instantiated into assembly here
    1 |         msr sctlr_el1, w8
      |                        ^

+++ arch/arm64/include/asm/sysreg.h
@@ -253,14 +253,14 @@ asm(
 "	.endm\n"
 );
 
-static inline void config_sctlr_el1(u32 clear, u32 set)
+static inline void config_sctlr_el1(u64 clear, u64 set)
 {
-	u32 val;
+    u64 val;
 
-	asm volatile("mrs %0, sctlr_el1" : "=r" (val));
-	val &= ~clear;
-	val |= set;
-	asm volatile("msr sctlr_el1, %0" : : "r" (val));
+    asm volatile("mrs %0, sctlr_el1" : "=r" (val));
+    val &= ~clear;
+    val |= set;
+    asm volatile("msr sctlr_el1, %0" : : "r" (val));
 }
 
 /*
